test: fixed failing test_in_out.py::test_hide_show due to teeardown function
`test_in_out.py::test_hide_show` was failing after my changes due to the teardown function calling `.stop()` on a spinner that somehow already had its output stream closed. This may indicate a problem, but I put `.stop()` in a try/except block since it seems inappropriate for a test to fail because a teardown function raises an exception.